unused: remove uses of typemap, deduplicate pointers
In 41bedbdbf9e6154d6edae49010d46bd375b076f2 we removed all uses of
typemap.Map because we no longer needed to deduplicate types. In
9680b434791dd7af8abb706f69b31c2270ac8c3a we reverted that decision
because we still had to deduplicate pointer types, because
types.NewPointer(T) != types.NewPointer(T).

At the time, reverting a commit was the easier choice. However, due to
type parameters in Go 1.18, the implementation of types.Identical, which
our map depended on, has gotten a lot harder to vendor and modify, due
to the use of unexported APIs for dealing with type parameters.

Thus we once again remove uses of typemap, and this time we explicitly
handle pointers by partially deduplicating them, and adding edges where
necessary to treat separate pointers as identical.